+++ docs/source/conf.py
@@ -41,7 +41,7 @@ extensions = [
     "sphinx_contributors",
     "sphinx.ext.intersphinx",
     "sphinxext.opengraph",
-    "sphinx_visualised",
+    "sphinx_visualized",
 ]
 
 # Add any paths that contain templates here, relative to this directory.