+++ sphinx_visualized/__init__.py
@@ -10,8 +10,9 @@ import shutil
 from collections import Counter
 from pathlib import Path
 from docutils import nodes as docutils_nodes
+from multiprocessing import Manager, Queue
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 
 def setup(app):
@@ -21,7 +22,7 @@ def setup(app):
 
     return {
         "version": __version__,
-        "parallel_read_safe": False,
+        "parallel_read_safe": True,
         "parallel_write_safe": True,
     }
 
@@ -33,17 +34,16 @@ def create_objects(app):
     builder = getattr(app, "builder", None)
     if builder is None:
         return
-    builder.env.app.nodes = [] # a list of nodes and their metadata
-    builder.env.app.links = [] # a list of links between pages
-    builder.env.app.pages = [] # an index of page names
-    builder.env.app.groups = [] # an index of page groups
+
+    manager = Manager()
+    builder.env.app.pages = manager.dict() # an index of page names
+    builder.env.app.references = manager.Queue() # a queue of every internal reference made between pages
 
 
 def get_links(app, doctree, docname):
     """
     Gather internal link connections
     """
-    references = [] # a list of every internal reference made between pages
 
     #TODO handle troctree entries?
     #TODO get targets
@@ -55,74 +55,22 @@ def get_links(app, doctree, docname):
         if node.tagname == 'reference' and node.get('internal') and node.get('refuri'):
             # calulate path of the referenced page in relation to docname
             ref = node.attributes['refuri'].split("#")[0]
-            absolute_ref = os.path.abspath(os.path.join(os.path.dirname(f"/{docname}.html"), ref))[1:-5]
+            refname = os.path.abspath(os.path.join(os.path.dirname(f"/{docname}.html"), ref))[1:-5]
 
             #TODO some how get ref/doc/term for type?
             # add each link as an individual reference
-            references.append((f"/{docname}.html", f"/{absolute_ref}.html", "ref"))
+            app.env.app.references.put((f"/{docname}.html", f"/{refname}.html", "ref"))
 
-            #TODO turn to function
-            # a group is the name of the top level directory
-            docname_group = f"/{docname}.html".split('/')[1]
-            if not docname_group in app.env.app.groups:
-                app.env.app.groups.append(docname_group)
+            docname_page = f"/{docname}.html"
+            app.env.app.pages[docname_page] = True
 
-            absolute_ref_group = f"/{absolute_ref}.html".split('/')[1]
-            if not absolute_ref_group in app.env.app.groups:
-                app.env.app.groups.append(absolute_ref_group)
-
-            # add to index of page names
-            if not f"/{docname}.html" in app.env.app.pages:
-                app.env.app.pages.append(f"/{docname}.html")
-
-            if not f"/{absolute_ref}.html" in app.env.app.pages:
-                app.env.app.pages.append(f"/{absolute_ref}.html")
-
-            # check if node exists based on docname's id
-            if not any(d.get("id") == app.env.app.pages.index(f"/{docname}.html") for d in app.env.app.nodes):
-                if app.env.titles.get(docname):
-                    title = app.env.titles.get(docname).astext()
-                else:
-                    title = f"/{docname}.html"
-
-                app.env.app.nodes.append({
-                    "id": app.env.app.pages.index(f"/{docname}.html"),
-                    "group": app.env.app.groups.index(docname_group),
-                    "label": title,
-                    "path": f"../../../{docname}.html",
-                    "level": 1
-                })
-
-            # check if node exists based on absolute_ref's id
-            if not any(d.get("id") == app.env.app.pages.index(f"/{absolute_ref}.html") for d in app.env.app.nodes):
-                if app.env.titles.get(absolute_ref):
-                    title = app.env.titles.get(absolute_ref).astext()
-                else:
-                    title = f"/{absolute_ref}.html"
-
-                app.env.app.nodes.append({
-                    "id": app.env.app.pages.index(f"/{absolute_ref}.html"),
-                    "group": app.env.app.groups.index(absolute_ref_group),
-                    "label": title,
-                    "path": f"../../../{absolute_ref}.html",
-                    "level": 1
-                })
-
-    # create object that links references between pages
-    references_counts = Counter(references)
-    for ref, count in references_counts.items():
-        #TODO check if unique else strength++
-        app.env.app.links.append({
-            "target": app.env.app.pages.index(ref[1]),
-            "source": app.env.app.pages.index(ref[0]),
-            "strength": 1,
-            "type": ref[2],
-        })
+            refname_page = f"/{refname}.html"
+            app.env.app.pages[refname_page] = True
 
 
 def build_toctree_hierarchy(app):
     """
-    Take toctree_includes and build the hierarchy while gathering page metadata.
+    Take toctree_includes and build the document hierarchy while gathering page metadata.
     """
     node_map = {}
     data = app.env.toctree_includes
@@ -153,6 +101,9 @@ def create_json(app, exception):
     """
     Create and copy static files for visualizations
     """
+    page_list = list(app.env.app.pages.keys()) # list of pages with references
+
+    # create directory in _static and over static assets
     os.makedirs(Path(app.outdir) / "_static" / "sphinx-visualized", exist_ok=True)
     if version.parse(sphinx.__version__) >= version.parse("8.0.0"):
         # Use the 'force' argument if it's available
@@ -169,13 +120,43 @@ def create_json(app, exception):
             os.path.join(app.builder.outdir, '_static', "sphinx-visualized"),
         )
 
+    # convert pages and groups to lists
+    nodes = [] # a list of nodes and their metadata
+    for page in page_list:
+        if app.env.titles.get(page[1:-5]):
+            title = app.env.titles.get(page[1:-5]).astext()
+        else:
+            title = page
+
+        nodes.append({
+            "id": page_list.index(page),
+            "label": title,
+            "path": f"../../..{page}",
+        })
+
+    # convert queue to list
+    reference_list = []
+    while not app.env.app.references.empty():
+        reference_list.append(app.env.app.references.get())
+
+    # create object that links references between pages
+    links = [] # a list of links between pages
+    references_counts = Counter(reference_list)
+    for ref, count in references_counts.items():
+        links.append({
+            "target": page_list.index(ref[1]),
+            "source": page_list.index(ref[0]),
+            "strength": count,
+            "type": ref[2],
+        })
+
     filename = Path(app.outdir) / "_static" / "sphinx-visualized" / "js" / "links.js"
     with open(filename, "w") as json_file:
-        json_file.write(f'var links_data = {json.dumps(app.env.app.links, indent=4)};')
+        json_file.write(f'var links_data = {json.dumps(links, indent=4)};')
 
     filename = Path(app.outdir) / "_static" / "sphinx-visualized" / "js" / "nodes.js"
     with open(filename, "w") as json_file:
-        json_file.write(f'var nodes_data = {json.dumps(app.env.app.nodes, indent=4)};')
+        json_file.write(f'var nodes_data = {json.dumps(nodes, indent=4)};')
 
     filename = Path(app.outdir) / "_static" / "sphinx-visualized" / "js" / "toctree.js"
     with open(filename, "w") as json_file: