+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 0.3.0
 -----
 
+*Release date: TBD*
+
 - Make toctree graph collapsible
 - Show the reference directionality in link graph
 - Consolidate files to single location
@@ -12,26 +14,36 @@ Changelog
 0.2.3
 -----
 
+*Release date: 2024-10-22*
+
 - Fix absolute_ref syntax so page titles are looked up correctly
 
 0.2.2
 -----
+
+*Release date: 2024-10-21*
 
 - Change json to js for data files
 
 0.2.1
 -----
 
+*Release date: 2024-10-17*
+
 - Fix MANIFEST
 
 0.2.0
 -----
+
+*Release date: 2024-10-17*
 
 - Rename extension to ``sphinx-visualised``
 - Add toctree graph
 
 0.1.0
 -----
+
+*Release date: 2024-08-26*
 
 - Initial release