+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 - Show the reference directionality in link graph
 - Consolidate files to single location
 - Support older versions of sphinx.util.fileutil.copy_asset()
+- Rename extension to ``sphinx-visualized``
 
 0.2.3
 -----